requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu117
# torch==2.0.1
# torchvision==0.15.2
packaging
sentencepiece
einops==0.4.1
fastapi==0.100.1
markdown2==2.4.10
numpy
openai==0.27.8
opencv_python
Pillow==9.4.0
pycocotools==2.0.6
ray==2.6.1
Requests==2.31.0
shortuuid==1.0.11
tqdm
transformers==4.36.2
uvicorn==0.23.2
scipy==1.11.2
bitsandbytes==0.41.1
timm
blobfile
mypy
pytest
requests
tensorboardX
ftfy
pyarrow
torchmetrics==0.7.3
deepspeed
pycocoevalcap
torchscale==0.2.0
hydra-core
accelerate
